fix: files renamed following better practices. Imports modified
