+++ api-proxy-server/src/routes/clubs.route.ts
