+++ api-proxy-server/src/routes/index.ts
@@ -1,3 +1,3 @@
-export * from "./player.route.js";
-export * from "./club.routes.js";
-export * from "./ranking.route.js";
+export * from "./players.route.js";
+export * from "./clubs.route.js";
+export * from "./rankings.route.js";
