+++ api-proxy-server/src/routes/rankings.route.ts
