+++ api-proxy-server/src/services/brawlstars.api.ts
