+++ api-proxy-server/src/types/client-error.ts
