+++ api-proxy-server/src/types/index.ts
@@ -1,8 +1,8 @@
-export * from "./clientError.js";
+export * from "./client-error.js";
 export * from "./club.js";
-export * from "./playerBattlelog.js";
-export * from "./playerInfo.js";
-export * from "./rankingGlobalBrawlers.js";
-export * from "./rankingGlobalClubs.js";
-export * from "./rankingGlobalPlayers.js";
+export * from "./player-battlelog.js";
+export * from "./player-info.js";
+export * from "./ranking-global-brawlers.js";
+export * from "./ranking-global-clubs.js";
+export * from "./ranking-global-players.js";
 export * from "./shared.js";
