+++ api-proxy-server/src/types/player-info.ts
