+++ api-proxy-server/src/types/ranking-global-clubs.ts
