+++ api-proxy-server/src/types/ranking-global-players.ts
