+++ frontend/src/App.tsx
@@ -1,5 +1,5 @@
 function App() {
-  return;
+  return <></>;
 }
 
 export default App;
