+++ frontend/src/api/brawlapi/client/brawl-api-client.ts
@@ -1,14 +1,14 @@
 import type { AxiosInstance } from "axios";
 import axios from "axios";
 
-export const BASE_URL = "https://api.brawlify.com/v1";
+export const BRAWLAPI_BASE_URL = "https://api.brawlify.com/v1";
 
 export class BrawlApiAxiosInterceptor {
   private axiosInstance: AxiosInstance;
 
   constructor() {
     this.axiosInstance = axios.create({
-      baseURL: BASE_URL,
+      baseURL: BRAWLAPI_BASE_URL,
       timeout: 5000,
       headers: {
         "Content-Type": "application/json",
